PLD_FinalProject_Group#_ParkinUP/requirements.txt
opencv-python
Pillow
pytesseract
